feat(apiserver): extract the API server into an importable package
Moves the UI's REST backend from cmd/api (a 2,400-line main package) to
pkg/dtrules/apiserver so `dtrules edit` can embed it; cmd/api becomes a
flag-parsing shim. First installment of the #757 SDK extraction. The load
parity test for #757 moves with the code.

Beyond the move, the server accumulated this session's correctness work:

- Parse the modern XML surface the old structs predated: <condition_dsl>/
  <action_dsl>/<context_dsl>, structured <initial_action> blocks (were
  silently dropped), and canonical policy statements
  (policy_statements>policy_statement with policy_description /
  policy_statement_postfix element names — broken for every format).
- Replace the entity/table maps with slices preserving authored document
  order, ending random list order on every load (Go map iteration).
  Name lookup is a linear-scan helper; source files live on the elements.
- Save through the canonical emitters (merge-on-save): parse the existing
  file, apply only API-editable fields, emit via excel.WriteXML /
  EDD WriteXML — preserving DSL/postfix pairs, collect/question metadata,
  and legacy tag forms the API model doesn't carry, and picking up
  provenance + number backfill. The old writers rewrote every table from
  every file through a lossy legacy schema.
- Entity numbers served and persisted; table lists sorted by
  TABLE_NUMBER; renumber-with-shift semantics (colliding number shifts
  everything at/after it down by 100); /api/dt/reorder and
  /api/edd/reorder renumber to match drag-and-drop order.
- /api/browse directory listing for the UI's project picker, confined by
  --project-root via the same traversal-safe validation as project open.
- /api/project/current so a fresh browser session adopts the project the
  server booted with.
- Config.ReadOnly: reject every mutating request and file browsing
  (reads, compilation, and execution stay available) for publishing
  rules for review.
